SKILLSWAP/SkillSwapHub/ODOO/SkillSwapHub/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
authlib==1.3.0
requests==2.31.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
